- First version of multi-threaded test. Only checks controller SPI until   now. - Implemented locking in SPIviaPCIe
git-svn-id: https://svnsrv.desy.de/desy/mtca4u/MotorDriverCard/branches/StepperMotor@400 5f601a5c-88b9-47f3-b851-818f6741d00d

+++ src/MotorDriverCardImpl.cc
@@ -159,7 +159,7 @@ namespace mtca4u{
     uint32_t maxFirmwareVersion =  (MINIMAL_FIRMWARE_VERSION & 0xFF000000) | 0x00FFFFFF;
 
     // check the version
-    if ( (static_cast<uint32_t>(firmwareVersion) <  MINIMAL_FIRMWARE_VERSION) || 
+    if ( (static_cast<uint32_t>(firmwareVersion) <  static_cast<uint32_t>(MINIMAL_FIRMWARE_VERSION)) || 
 	 (static_cast<uint32_t>(firmwareVersion) >  maxFirmwareVersion) ){
       std::stringstream errorMessage;
       errorMessage <<  "MotorDriverCardImpl detected wrong firmware version. "
